Migrate OpDescription functionality to Infos
